WIP: moved http util functions into a class
