+++ functions/src/HttpUtil.test.ts
@@ -4,14 +4,11 @@ import {mockReq} from 'sinon-express-mock'
 chai.use(sinonChai)
 import 'mocha'
 
-import {
-  getQueryStringFromArray,
-  getNextActionUrl,
-  getCurrentUrl,
-  parseQueryStringToArray
-} from './util'
+import {HttpRequestUtil} from './HttpRequestUtil'
+
 const expect = chai.expect
 
+const httpUtil = new HttpRequestUtil()
 const request = mockReq({
   protocol: 'http',
   originalUrl: '/voice',
@@ -25,21 +22,21 @@ describe('util function tests', () => {
   afterEach(() => {})
 
   it('test getQueryString from array of numbers is encdoed csv string', () => {
-    const str = getQueryStringFromArray(['+44111111', '+44222222'])
+    const str = httpUtil.getQueryStringFromArray(['+44111111', '+44222222'])
     expect('%2B44111111%2C%2B44222222').to.equal(str)
   })
   it('test getUrl ', () => {
-    const str = getCurrentUrl(request)
+    const str = httpUtil.getCurrentUrl(request)
     console.log(str)
     expect('http://localhost/voice').to.equal(str)
   })
   it('test getNextActionUrl appends number to query string', () => {
-    const str = getNextActionUrl(request, '+44999999')
+    const str = httpUtil.getNextActionUrl(request, '+44999999')
     console.log(str)
     expect('http://localhost/voice?numbersTried=%2B44999999').to.equal(str)
   })
   it('test getQueryString ', () => {
-    const arr = parseQueryStringToArray(
+    const arr = httpUtil.parseQueryStringToArray(
       mockReq({
         protocol: 'http',
         originalUrl: '/voice',
