--- functions/src/test.ts
@@ -1,13 +1,0 @@
-import * as chai from 'chai'
-import 'mocha'
-const expect = chai.expect
-
-describe('Sample tests', () => {
-  before(() => {
-    expect(true).to.be.true
-  })
-
-  afterEach(() => {})
-
-  it('test something', async () => {})
-})
